+++ src/main/java/acceso_datos/consultas_bd/IControladorCBDRecetasCooker.java
@@ -34,5 +34,4 @@ public interface IControladorCBDRecetasCooker {
     public Chef consultaRecetaXChef (BigInteger idReceta) throws  SQLException;
     public Categoria consultaCategoria(int idCategoria) throws SQLException;
     public List<Categoria> consultarCategorias() throws SQLException;
-    public boolean calificarChef (BigInteger idChef, BigInteger idCooker, int calificacion) throws SQLException;
 }
