+++ src/main/java/acceso_datos/persistencia_bd/IControladorPBDRecetasCooker.java
@@ -9,4 +9,5 @@ public interface IControladorPBDRecetasCooker {
     public boolean crearListaFavoritosConReceta(DTOListaFavoritos listaFavoritos) throws SQLException;
     public boolean crearListaFavoritos(DTOListaFavoritos listaFavoritos) throws SQLException;
     public boolean insertarRecetaListaFavoritos(BigInteger idreceta, int idlista, BigInteger idusuario) throws SQLException;
+    public boolean calificarChef (BigInteger idChef, BigInteger idCooker, int calificacion) throws SQLException;
 }
